+++ tiger-testenv-mgr/src/test/java/de/gematik/test/tiger/testenvmgr/config/ForwardProxyInfoTest.java
@@ -11,6 +11,7 @@ import static org.awaitility.Awaitility.await;
 
 import de.gematik.rbellogger.data.facet.RbelHttpRequestFacet;
 import de.gematik.test.tiger.common.config.TigerGlobalConfiguration;
+import de.gematik.test.tiger.common.pki.TigerConfigurationPkiIdentity;
 import de.gematik.test.tiger.proxy.TigerProxy;
 import de.gematik.test.tiger.testenvmgr.TigerTestEnvMgr;
 import de.gematik.test.tiger.testenvmgr.junit.TigerTest;
@@ -20,14 +21,20 @@ import java.io.IOException;
 import java.io.OutputStream;
 import java.net.ServerSocket;
 import java.net.Socket;
+import java.security.cert.CertificateException;
+import java.security.cert.X509Certificate;
+import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
+import javax.net.ssl.*;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Test;
 
 @Slf4j
 class ForwardProxyInfoTest {
+
+  private AtomicBoolean shouldServerRun = new AtomicBoolean(true);
 
   @SneakyThrows
   @Test
@@ -104,24 +111,16 @@ class ForwardProxyInfoTest {
       tigerYaml =
           """
         tigerProxy:
-          forwardToProxy:
-            hostname: 127.0.0.1
-            port: ${free.port.20}
         servers:
-          someProxyServer:
-            type: tigerProxy
-            tigerProxyCfg:
-              adminPort: ${free.port.11}
-              proxyPort: ${free.port.21}
           virtualExternalServer:
             type: externalUrl
             source:
-              - http://localhost:${free.port.31}
+              - http://localhost:${free.port.30}
         """,
       skipEnvironmentSetup = true)
-  void localSourceNonHttpAndForwardProxyConfigured_shouldStillStartUp(TigerTestEnvMgr envMgr) {
+  void failingHttpsServer_shouldStillStartUp(TigerTestEnvMgr envMgr) {
     AtomicBoolean shouldServerRun = new AtomicBoolean(true);
-    final Integer port = TigerGlobalConfiguration.readIntegerOptional("free.port.31").orElseThrow();
+    final Integer port = TigerGlobalConfiguration.readIntegerOptional("free.port.30").orElseThrow();
     try (ServerSocket serverSocket = new ServerSocket(port)) {
       final Thread serverThread =
           new Thread(
@@ -130,7 +129,7 @@ class ForwardProxyInfoTest {
                   try {
                     Socket socket = serverSocket.accept();
                     OutputStream out = socket.getOutputStream();
-                    out.write("fdsa".getBytes());
+                    out.write("pong".getBytes());
                     out.close();
                     socket.close();
                   } catch (IOException e) {
@@ -143,6 +142,94 @@ class ForwardProxyInfoTest {
     } finally {
       shouldServerRun.set(false);
     }
+  }
+
+  @SneakyThrows
+  @Test
+  @TigerTest(
+      tigerYaml =
+          """
+        tigerProxy:
+          forwardToProxy:
+            hostname: 127.0.0.1
+            port: ${free.port.20}
+        servers:
+          someProxyServer:
+            type: tigerProxy
+            tigerProxyCfg:
+              adminPort: ${free.port.11}
+              proxyPort: ${free.port.21}
+          virtualExternalServer:
+            type: externalUrl
+            source:
+              - https://localhost:${free.port.31}
+        """,
+      skipEnvironmentSetup = true)
+  void localSourceNonHttpAndForwardProxyConfigured_shouldStillStartUp(TigerTestEnvMgr envMgr) {
+    AtomicBoolean shouldServerRun = new AtomicBoolean(true);
+    final Integer port = TigerGlobalConfiguration.readIntegerOptional("free.port.31").orElseThrow();
+    try {
+      startDistrustingSslServer(port);
+      assertThatNoException().isThrownBy(envMgr::setUpEnvironment);
+    } finally {
+      shouldServerRun.set(false);
+    }
+  }
+
+  @SneakyThrows
+  public int startDistrustingSslServer(Integer port) {
+    var threadPool = Executors.newCachedThreadPool();
+    SSLContext sslContext = getSSLContext();
+    SSLServerSocketFactory ssf = sslContext.getServerSocketFactory();
+    SSLServerSocket serverSocket = (SSLServerSocket) ssf.createServerSocket(port);
+    serverSocket.setNeedClientAuth(true);
+
+    threadPool.execute(
+        () -> {
+          while (shouldServerRun.get()) {
+            try {
+              Socket socket = serverSocket.accept();
+              socket.close();
+            } catch (IOException e) {
+              // swallow
+            }
+          }
+        });
+
+    return serverSocket.getLocalPort();
+  }
+
+  protected SSLContext getSSLContext() throws Exception {
+    SSLContext sslContext = SSLContext.getInstance("TLS");
+    final TigerConfigurationPkiIdentity serverCert =
+        new TigerConfigurationPkiIdentity(
+            "../tiger-proxy/src/test/resources/eccStoreWithChain.jks;gematik");
+    // Set up key manager factory to use our key store
+    KeyManagerFactory kmf = KeyManagerFactory.getInstance(KeyManagerFactory.getDefaultAlgorithm());
+    kmf.init(serverCert.toKeyStoreWithPassword("00"), "00".toCharArray());
+
+    // Initialize the SSLContext to work with our key managers.
+    final X509TrustManager x509TrustManager =
+        new X509TrustManager() {
+          @Override
+          public void checkClientTrusted(X509Certificate[] chain, String authType)
+              throws CertificateException {
+            throw new CertificateException("go fish");
+          }
+
+          @Override
+          public void checkServerTrusted(X509Certificate[] chain, String authType) {
+            // swallow
+          }
+
+          @Override
+          public X509Certificate[] getAcceptedIssuers() {
+            return new X509Certificate[] {};
+          }
+        };
+    sslContext.init(kmf.getKeyManagers(), new X509TrustManager[] {x509TrustManager}, null);
+
+    return sslContext;
   }
 
   @SneakyThrows