+++ ReleaseNotes.md
@@ -7,6 +7,7 @@
 * TGR-1125: Tiger-Zion: Fixed assignment-bugs
 * TGR-1183: fixed an issue where TigerConfigurationKeys were wrongly ignoring parts of the key when they add a repeated
   subkey.
+* TGR-1186: Fixed healthcheck-issues with TLS servers
 * TGR-1180: fixed position of close button in the configuration editor in the WorkflowUi
 * TGR-1143: in case of missing dependencies for a server type the exception is more clearly stating that there is a
   dependency missing